Tidy up according to PEP8.
