+++ src/mapclient/view/managers/options/optionsdialog.py
@@ -10,7 +10,7 @@ from PySide6 import QtWidgets
 from mapclient.view.managers.options.ui.ui_optionsdialog import Ui_OptionsDialog
 
 from mapclient.core.checks import WizardToolChecks, VCSChecks
-from mapclient.view.syntaxhighlighter import SyntaxHighlighter
+from mapclient.view.utils import SyntaxHighlighter
 from mapclient.settings.definitions import VIRTUAL_ENVIRONMENT_STRING, \
     WIZARD_TOOL_STRING, PMR_TOOL_STRING, INTERNAL_WORKFLOWS_AVAILABLE
